+++ app/components/Form/index.ts
@@ -1,0 +1,1 @@
+export { Form, type FormProps } from "./Form.tsx"
